정보게시판, 마이페이지, 회원가입 수정 및 CSS 적용

+++ src/main/webapp/WEB-INF/views/service_login/withdrawal.jsp
@@ -5,40 +5,53 @@
 <head>
 <meta charset="UTF-8">
 <title>회원 탈퇴</title>
-<style type="text/css">
-	.form { margin-top: 350px;}
-	a { text-decoration: none; color: black; }
-	table { background: white; padding: 30px; }
-</style>
 <script src="resources/jquery-3.5.1.min.js"></script>
 <script type="text/javascript">
+	function chk_loginUser() {
+		if ('${loginUser}' == "") {
+			alert("로그인 후 사용 가능합니다.")
+			location.href="login"
+		}else {
+			console.log("로그인 확인 성공")
+		}
+	}
 	function chk_pwd() {
 		var id = '${loginUser.id}'
 		var pwd = $('input[name=pwd]').val()
-		var form = {
-			id : id,
-			pwd : pwd
-		}
-		$.ajax({
-			url : "chk_dbpwd",
-			type : "POST",
-			data : form,
-			success : function(result) {
-				if (result == "비밀번호가 틀렸습니다!") {
-					alert("비밀번호가 틀립니다.\n다시 입력해주세요.")
-					$('input[name=pwd]').val("")
-					$('input[name=pwd]').focus
-				} else {
-					console.log("성공")
-					delete_User()
-				}
-			},
-			error : function(request, status, error) {
-				console.log("실패")
-				alert("code:" + request.status + "\n" + "message:"
-						+ request.responseText + "\n" + "error:" + error);
+		if(pwd=="") {
+			alert("비밀번호를 입력하세요.")
+			$('input[name=pwd]').focus()
+		}else {
+			var form = {
+				id : id,
+				pwd : pwd
 			}
-		})
+			$.ajax({
+				url : "chk_dbpwd",
+				type : "POST",
+				data : form,
+				success : function(result) {
+					if (result == "비밀번호가 틀렸습니다!") {
+						alert("비밀번호가 틀립니다.\n다시 입력해주세요.")
+						$('input[name=pwd]').val("")
+						$('input[name=pwd]').focus()
+					} else {
+						console.log("성공")
+						var con_test = confirm("정말로 탈퇴하시겠습니까?")
+						if(con_test) {
+							delete_User()
+						} else {
+							alert("회원 탈퇴를 취소하였습니다.")
+						}
+					}
+				},
+				error : function(request, status, error) {
+					console.log("실패")
+					alert("code:" + request.status + "\n" + "message:"
+							+ request.responseText + "\n" + "error:" + error);
+				}
+			})
+		}
 	}
 	function delete_User() {
 		var id = '${loginUser.id}'
@@ -66,23 +79,55 @@
 	}
 </script>
 </head>
-<body>
-	<%@ include file="../defualt/header.jsp" %>
-	<form action="" class="form">
-		<h1>회원 탈퇴</h1>
-			회원 탈퇴 시 Trevelers 사이트 내에 저장된 모든 정보가 삭제되며, 이후 복구가 불가합니다.<br>
-			본인 확인을 위해 비밀번호를 입력하고, 이메일 인증을 진행해주세요.<br>
-			<table border="1">	
-				<caption><h3>회원 탈퇴를 요청한 아이디</h3></caption>
-				<tr>
-					<th>아이디</th><th>닉네임</th><th>이메일</th>
-				</tr>
-				<tr>
-					<td>${loginUser.id}</td><td>${loginUser.nick}</td><td>${loginUser.email}</td>
-				</tr>
-			</table>
-			<input type="password" name="pwdok"><input type="button" value="회원탈퇴" onclick="delete_User()"><br><br><br>
-		</form>
+<body class="is-preload" onload="chk_loginUser()">
+	<%@ include file="../defualt/header.jsp"%>
+	<!-- Page Wrapper -->
+	<div id="page-wrapper">
+		<!-- Main -->
+		<article id="main">
+			<section class="wrapper style5">
+				<div class="inner">
+					<section>
+						<h2>My Page</h2>
+						<div class="row">
+							<div class="col-6 col-12-medium" style="width:200px;">
+								<ul class="alt">
+									<li><a href="mypage">내정보</a></li>
+									<li><a href="travelDiary">여행수첩</a></li>
+									<li><a href="mypage">메이트 신청 현황</a></li>
+									<li><a href="mypage">작성한 글 관리</a></li>
+									<li><a href="chk_pwd?page=change_userinfo">회원정보 수정</a></li>
+									<li><a href="chk_pwd?page=change_pwd">비밀번호 수정</a></li>
+									<li><a href="withdrawal">회원탈퇴</a></li>
+								</ul>
+							</div>
+							<div class="col-6 col-12-medium" style="margin-left: 50px;width:700px;">
+								<h2>회원 탈퇴</h2>
+								회원 탈퇴 시 Trevelers 사이트 내에 저장된 모든 개인 정보가 삭제되며,<br>
+								이후 계정 복구가 불가합니다.<br>
+								본인 확인을 위해 비밀번호를 입력하고, 이메일 인증을 진행해주세요.<br><br>
+								<table border="1">	
+									<caption><h3>회원 탈퇴를 요청한 아이디</h3></caption>
+									<tr style="text-align: center;">
+										<td>아이디</td>
+										<td>닉네임</td>
+										<td>이메일</td>
+									</tr>
+									<tr style="text-align: right;" >
+										<td>${loginUser.id}</td>
+										<td>${loginUser.nick}</td>
+										<td>${loginUser.email}</td>
+									</tr>
+								</table>
+								<input type="password" name="pwd" autofocus="autofocus" placeholder="비밀번호를 입력하세요."><br>
+								<input style="text-align: right;" type="button" value="회원탈퇴" onclick="chk_pwd()">
+							</div>
+						</div>
+					</section>
+				</div>
+			</section>
+		</article>
+	</div>
 	<%@ include file="../defualt/footer.jsp" %>
 </body>
 </html>
